+++ tsconfig-base.json
@@ -1,0 +1,34 @@
+{
+	"compilerOptions": {
+		"allowJs": true,
+		"allowSyntheticDefaultImports": true,
+		"baseUrl": "src",
+		"declaration": true,
+		"esModuleInterop": true,
+		"inlineSourceMap": false,
+		"lib": [
+			"esnext"
+		],
+		"listEmittedFiles": false,
+		"listFiles": false,
+		"moduleResolution": "node",
+		"noFallthroughCasesInSwitch": true,
+		"pretty": true,
+		"resolveJsonModule": true,
+		"rootDir": "src",
+		"skipLibCheck": true,
+		"strict": true,
+		"traceResolution": false,
+		"types": [
+			"node"
+		]
+	},
+	"compileOnSave": false,
+	"exclude": [
+		"node_modules",
+		"dist"
+	],
+	"include": [
+		"src"
+	]
+}
