+++ tsconfig-cjs.json
@@ -1,0 +1,8 @@
+{
+	"extends": "./tsconfig-base.json",
+	"compilerOptions": {
+		"module": "commonjs",
+		"outDir": "dist/cjs",
+		"target": "es2015"
+	}
+}
